Tüm ana dosyalar srcde olur, config hariç normali böyle bunun. Ayrıca config artık gizli
