+++ config.example.json
@@ -1,0 +1,5 @@
+{
+    "prefix": "!",
+    "token": "TOKEN",
+    "mongo": "MONGODB"
+}
