+++ package.json
@@ -2,6 +2,7 @@
     "scripts": {
         "start": "node ."
     },
+    "main": "src/index.js",
     "dependencies": {
         "@discordjs/rest": "^1.6.0",
         "discord.js": "^14.11.0",
